Adde the distance based moves generator.
This is still work in progress. The problem of
forcing (escaping) moves isn't solved: if the
defender makes a move on the edge of the
r zone, the attacker must block it and it's not
clear how to detect such moves. Without this,
the r zone quickly spans across the entire board.