Implmeneted a simplified lambda-search.
The idea is to stop searching once the target
gets to make more liberties than a certain threshold.
Then if target cannot be captured with the current
threshold, it's increased and a new search is started.

This search doesn't remember what moves were
played by the previous search (aka "the shadow zone"
in the T. Thompsen's paper), so new searches try all
moves within the fully enclosed area. Intermediate
result are cached, though.