Use 64 bit hashes in tt
The expected number of collisions is
E = n*n/(2*m) where n is the number
of entries and m is 2**32 or 2**64,
depending on the key size.

     n      32 bit      64 bit
     --------------------------
     1M     0.11e3      2.71e-8
     2M     0.47e3      1.08e-7
     3M     1.05e3      2.43e-7
     4M     1.86e3      4.34e-7
     5M     2.91e3      6.78e-7

So with 32 bit hashes a few hundreds of
collisions in one tsumego is a sure thing,
while with 64 bit hases there will be one
collision per 10 million tsumegos.

This comes with a price: the solver now needs
265s to solve the 100 tsumegos, which is 1.1x
slower.